+++ src/Sylius/Bundle/CoreBundle/Controller/OrderController.php
@@ -11,116 +11,14 @@
 
 namespace Sylius\Bundle\CoreBundle\Controller;
 
-use Doctrine\ORM\EntityManager;
 use FOS\RestBundle\View\View;
-use Payum\Core\Registry\RegistryInterface;
-use Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration;
-use Sylius\Bundle\ResourceBundle\Controller\ResourceController;
-use Sylius\Component\Order\Context\CartContextInterface;
-use Sylius\Component\Order\Model\OrderInterface;
-use Sylius\Component\Order\SyliusCartEvents;
-use Sylius\Component\Resource\ResourceActions;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Component\EventDispatcher\GenericEvent;
-use Symfony\Component\HttpFoundation\RedirectResponse;
+use Sylius\Bundle\OrderBundle\Controller\OrderController as BaseOrderController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\HttpException;
 use Webmozart\Assert\Assert;
 
-class OrderController extends ResourceController
+class OrderController extends BaseOrderController
 {
-    /**
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function summaryAction(Request $request)
-    {
-        $configuration = $this->requestConfigurationFactory->create($this->metadata, $request);
-
-        $cart = $this->getCurrentCart();
-
-        if (!$configuration->isHtmlRequest()) {
-            return $this->viewHandler->handle($configuration, View::create($cart));
-        }
-
-        $form = $this->resourceFormFactory->create($configuration, $cart);
-
-        $view = View::create()
-            ->setTemplate($configuration->getTemplate('summary.html'))
-            ->setData([
-                'cart' => $cart,
-                'form' => $form->createView(),
-            ])
-        ;
-
-        return $this->viewHandler->handle($configuration, $view);
-    }
-
-    /**
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function saveAction(Request $request)
-    {
-        $configuration = $this->requestConfigurationFactory->create($this->metadata, $request);
-
-        $this->isGrantedOr403($configuration, ResourceActions::UPDATE);
-        $resource = $this->getCurrentCart();
-
-        $form = $this->resourceFormFactory->create($configuration, $resource);
-
-        if (in_array($request->getMethod(), ['POST', 'PUT', 'PATCH'], true) && $form->handleRequest($request)->isValid()) {
-            $resource = $form->getData();
-
-            $event = $this->eventDispatcher->dispatchPreEvent(ResourceActions::UPDATE, $configuration, $resource);
-
-            if ($event->isStopped() && !$configuration->isHtmlRequest()) {
-                throw new HttpException($event->getErrorCode(), $event->getMessage());
-            }
-            if ($event->isStopped()) {
-                $this->flashHelper->addFlashFromEvent($configuration, $event);
-
-                return $this->redirectHandler->redirectToResource($configuration, $resource);
-            }
-
-            if ($configuration->hasStateMachine()) {
-                $this->stateMachine->apply($configuration, $resource);
-            }
-
-            $this->eventDispatcher->dispatchPostEvent(ResourceActions::UPDATE, $configuration, $resource);
-
-            if (!$configuration->isHtmlRequest()) {
-                return $this->viewHandler->handle($configuration, View::create(null, Response::HTTP_NO_CONTENT));
-            }
-
-            $this->getEventDispatcher()->dispatch(SyliusCartEvents::CART_CHANGE, new GenericEvent($resource));
-            $this->manager->flush();
-
-            $this->flashHelper->addSuccessFlash($configuration, ResourceActions::UPDATE, $resource);
-
-            return $this->redirectHandler->redirectToResource($configuration, $resource);
-        }
-
-        if (!$configuration->isHtmlRequest()) {
-            return $this->viewHandler->handle($configuration, View::create($form, Response::HTTP_BAD_REQUEST));
-        }
-
-        $view = View::create()
-            ->setData([
-                'configuration' => $configuration,
-                $this->metadata->getName() => $resource,
-                'form' => $form->createView(),
-                'cart' => $resource,
-            ])
-            ->setTemplate($configuration->getTemplate(ResourceActions::UPDATE . '.html'))
-        ;
-
-        return $this->viewHandler->handle($configuration, $view);
-    }
-
     /**
      * @param Request $request
      *
@@ -154,90 +52,5 @@ class OrderController extends ResourceController
         ;
 
         return $this->viewHandler->handle($configuration, $view);
-    }
-
-    /**
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function clearAction(Request $request)
-    {
-        $configuration = $this->requestConfigurationFactory->create($this->metadata, $request);
-
-        $this->isGrantedOr403($configuration, ResourceActions::DELETE);
-        $resource = $this->getCurrentCart();
-
-        if ($configuration->isCsrfProtectionEnabled() && !$this->isCsrfTokenValid($resource->getId(), $request->get('_csrf_token'))) {
-            throw new HttpException(Response::HTTP_FORBIDDEN, 'Invalid csrf token.');
-        }
-
-        $event = $this->eventDispatcher->dispatchPreEvent(ResourceActions::DELETE, $configuration, $resource);
-
-        if ($event->isStopped() && !$configuration->isHtmlRequest()) {
-            throw new HttpException($event->getErrorCode(), $event->getMessage());
-        }
-        if ($event->isStopped()) {
-            $this->flashHelper->addFlashFromEvent($configuration, $event);
-
-            return $this->redirectHandler->redirectToIndex($configuration, $resource);
-        }
-
-        $this->repository->remove($resource);
-        $this->eventDispatcher->dispatchPostEvent(ResourceActions::DELETE, $configuration, $resource);
-
-        if (!$configuration->isHtmlRequest()) {
-            return $this->viewHandler->handle($configuration, View::create(null, Response::HTTP_NO_CONTENT));
-        }
-
-        $this->flashHelper->addSuccessFlash($configuration, ResourceActions::DELETE, $resource);
-
-        return $this->redirectHandler->redirectToIndex($configuration, $resource);
-    }
-
-    /**
-     * @param RequestConfiguration $configuration
-     *
-     * @return RedirectResponse
-     */
-    protected function redirectToCartSummary(RequestConfiguration $configuration)
-    {
-        if (null === $configuration->getParameters()->get('redirect')) {
-            return $this->redirectHandler->redirectToRoute($configuration, $this->getCartSummaryRoute());
-        }
-
-        return $this->redirectHandler->redirectToRoute($configuration, $configuration->getParameters()->get('redirect'));
-    }
-
-    /**
-     * @return string
-     */
-    protected function getCartSummaryRoute()
-    {
-        return 'sylius_cart_summary';
-    }
-
-    /**
-     * @return OrderInterface
-     */
-    protected function getCurrentCart()
-    {
-        return $this->getContext()->getCart();
-    }
-
-    /**
-     * @return CartContextInterface
-     */
-    protected function getContext()
-    {
-        return $this->get('sylius.context.cart');
-    }
-
-    /**
-     * @return EventDispatcherInterface
-     */
-    protected function getEventDispatcher()
-    {
-        return $this->container->get('event_dispatcher');
     }
 }
