+++ src/Sylius/Bundle/CoreBundle/Form/Extension/OrderTypeExtension.php
@@ -13,12 +13,8 @@ namespace Sylius\Bundle\CoreBundle\Form\Extension;
 
 use Sylius\Bundle\AddressingBundle\Form\Type\AddressType;
 use Sylius\Bundle\OrderBundle\Form\Type\OrderType;
-use Sylius\Bundle\PromotionBundle\Form\Type\PromotionCouponToCodeType;
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Form\FormInterface;
-use Symfony\Component\OptionsResolver\Options;
-use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
@@ -33,28 +29,7 @@ final class OrderTypeExtension extends AbstractTypeExtension
         $builder
             ->add('shippingAddress', AddressType::class)
             ->add('billingAddress', AddressType::class)
-            ->add('promotionCoupon', PromotionCouponToCodeType::class, [
-                'by_reference' => false,
-                'label' => 'sylius.form.cart.coupon',
-                'required' => false,
-            ])
         ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver->setNormalizer('validation_groups', function (Options $options, array $validationGroups) {
-            return function (FormInterface $form) use ($validationGroups) {
-                if ((bool) $form->get('promotionCoupon')->getNormData()) { // Validate the coupon if it was sent
-                    $validationGroups[] = 'sylius_promotion_coupon';
-                }
-
-                return $validationGroups;
-            };
-        });
     }
 
     /**
