+++ src/Sylius/Bundle/OrderBundle/Form/Type/CartItemType.php
@@ -1,0 +1,65 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sylius\Bundle\OrderBundle\Form\Type;
+
+use Sylius\Bundle\ResourceBundle\Form\Type\AbstractResourceType;
+use Symfony\Component\Form\DataMapperInterface;
+use Symfony\Component\Form\Extension\Core\Type\IntegerType;
+use Symfony\Component\Form\FormBuilderInterface;
+
+/**
+ * @author Paweł Jędrzejewski <pawel@sylius.org>
+ */
+class CartItemType extends AbstractResourceType
+{
+    /**
+     * @var DataMapperInterface
+     */
+    private $dataMapper;
+
+    /**
+     * @param string $dataClass
+     * @param array $validationGroups
+     * @param DataMapperInterface $dataMapper
+     */
+    public function __construct(
+        $dataClass,
+        array $validationGroups = [],
+        DataMapperInterface $dataMapper
+    ) {
+        parent::__construct($dataClass, $validationGroups);
+
+        $this->dataMapper = $dataMapper;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder
+            ->add('quantity', IntegerType::class, [
+                'attr' => ['min' => 1],
+                'label' => 'sylius.ui.quantity',
+            ])
+            ->setDataMapper($this->dataMapper)
+        ;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getBlockPrefix()
+    {
+        return 'sylius_cart_item';
+    }
+}
