+++ src/Sylius/Bundle/OrderBundle/Form/Type/CartType.php
@@ -13,13 +13,12 @@ namespace Sylius\Bundle\OrderBundle\Form\Type;
 
 use Sylius\Bundle\ResourceBundle\Form\Type\AbstractResourceType;
 use Symfony\Component\Form\Extension\Core\Type\CollectionType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
  */
-final class OrderType extends AbstractResourceType
+class CartType extends AbstractResourceType
 {
     /**
      * {@inheritdoc}
@@ -28,14 +27,11 @@ final class OrderType extends AbstractResourceType
     {
         $builder
             ->add('items', CollectionType::class, [
-                'entry_type' => OrderItemType::class,
+                'entry_type' => CartItemType::class,
                 'allow_add' => true,
                 'allow_delete' => true,
                 'by_reference' => false,
                 'label' => 'sylius.form.order.items',
-            ])
-            ->add('notes', TextareaType::class, [
-                'label' => 'sylius.ui.notes',
             ])
         ;
     }
@@ -45,6 +41,6 @@ final class OrderType extends AbstractResourceType
      */
     public function getBlockPrefix()
     {
-        return 'sylius_order';
+        return 'sylius_cart';
     }
 }
