+++ src/Sylius/Bundle/ShopBundle/Resources/views/Checkout/SelectShipping/_choice.html.twig
@@ -1,3 +1,5 @@
+{% import '@SyliusShop/Common/Macro/money.html.twig' as money %}
+
 <div class="item">
     <div class="field">
         <div class="ui radio checkbox">
@@ -14,7 +16,7 @@
     </div>
     <div class="extra">
         <div class="ui large right floated fee label">
-            {{ fee|sylius_price }}
+            {{ money.convertAndFormat(fee) }}
         </div>
     </div>
 </div>
